[pandas 2.0] Scaffolding for initializing pandas arrays from NumPy arrays
This relates to our discussions in the design documents.    * Non-
strided NumPy arrays (without missing data, some issues to address
here) are initialized with zero copy.   * Strided arrays are copied
Some more tests to write and implementation to do yet. In my opinion,
this code shouldn't be too smart -- if you need to examine the
incoming data to determine which values are null, that logic should be
elsewhere. This is responsible for constructing the array containers
with the right metadata.    Code review on Gerrit:
https://review.gerrithub.io/#/c/298548/

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #51 from wesm/arrays-from-numpy and squashes the following commits:

5f58fc7 [Wes McKinney] Instantiate NumericArray for floating point types
0663387 [Wes McKinney] Basic infrastructure for converting NumPy arrays to pandas arrays, strided and unstrided, unit tests.